test_edition
